+++ packages/appcd-plugin/src/tunnel-stream.js
@@ -1,5 +1,5 @@
 import util from 'util';
-
+import { redact } from 'appcd-util';
 import { Writable } from 'stream';
 
 /**
@@ -27,7 +27,7 @@ export default class TunnelStream extends Writable {
 		if (process.connected && typeof message === 'object') {
 			message.args = [
 				util.format(...message.args.map(a => {
-					return typeof a === 'string' ? a : util.inspect(scrub(a), { colors: true, depth: null });
+					return typeof a === 'string' ? a : util.inspect(redact(a), { colors: true, depth: null });
 				}))
 			];
 
@@ -38,35 +38,4 @@ export default class TunnelStream extends Writable {
 		}
 		cb();
 	}
-}
-
-/**
- * A lookup of various properties that must be redacted during log message serialization.
- * @type {Set}
- */
-const sensitiveProperties = new Set([
-	'clientsecret',
-	'password'
-]);
-
-/**
- * Deeply copies an object and scrubs any potentially sensitive data.
- *
- * @param {Object} src - The source object to copy from.
- * @returns {Object}
- */
-function scrub(src) {
-	if (Array.isArray(src)) {
-		return src.map(scrub);
-	}
-
-	if (src && typeof src === 'object') {
-		const dest = {};
-		for (const [ key, value ] of Object.entries(src)) {
-			dest[key] = sensitiveProperties.has(key.toLowerCase()) ? '<REDACTED>' : scrub(value);
-		}
-		return dest;
-	}
-
-	return src;
 }